fix: ssr and csr mismatch by disabling ssr
zag-js led to hydration issues. disabling ssr for now because generally easier architecture
