+++ source-code/website/src/pages/documentation/@id/index.page.server.tsx
@@ -1,26 +1,8 @@
-/**
- * -------------------------------------
- * Only runs server side because markdown files are
- * read from the filesystem.
- *
- * Be aware that routing to this page needs to happen server side.
- * -------------------------------------
- */
-
+import type { OnBeforeRender } from "@src/renderer/types.js";
 import { parseValidateAndRender } from "@src/services/markdoc/parseValidateAndRender.js";
-import { Header } from "./Header.jsx";
-import { Navigation } from "./Navigation.jsx";
-import { tableOfContent } from "./tableOfContent.js";
-import type { OnBeforeRender, PageHead } from "@src/renderer/types.js";
-import { Footer } from "./Footer.jsx";
+import type { PageProps } from "./index.page.jsx";
 
-export const Head: PageHead = () => {
-	return {
-		title: "Documentation",
-		description: "Documentation",
-	};
-};
-
+// should only run server side
 export const onBeforeRender: OnBeforeRender<PageProps> = async (
 	pageContext
 ) => {
@@ -44,28 +26,3 @@ export const onBeforeRender: OnBeforeRender<PageProps> = async (
 		};
 	}
 };
-
-type PageProps = {
-	markdown?: string;
-};
-
-export function Page(props: PageProps) {
-	return (
-		<>
-			<Header />
-			<div class="mx-auto max-w-screen-2xl  sm:px-6 md:px-0  ">
-				<div class=" flex ">
-					<Navigation sections={tableOfContent} />
-
-					{props.markdown ? (
-						<div class=" w-full prose mx-auto" innerHTML={props.markdown}></div>
-					) : (
-						<p>schade</p>
-					)}
-				</div>
-			</div>
-
-			<Footer />
-		</>
-	);
-}
